Add the possibility to use overlapping windows in cross-spectrum analysis
For this, we add a new option EP.noverlap which must be between 0 and nfft-1, and is typically set to nfft/2 (default behavior of cpsd)
The default value is 0 so as to have unchanged DIWASP behavior.
If noverlap is specified, nfft must be specified too so that we can check the consistency of the two values.